fix IGraphicsCairo SVG loading adding static storage for svg (mac)

+++ IPlug/IGraphics/IGraphicsCairo.cpp
@@ -203,8 +203,8 @@ void IGraphicsCairo::DrawSVG(ISVG& svg, const IRECT& dest, const IBlend* pBlend)
   cairo_rectangle(mContext, 0, 0, dest.W(), dest.H());
   cairo_clip(mContext);
 
-  double xScale = dest.W() / svg.W();
-  double yScale = dest.H() / svg.H();
+  double xScale = (double) dest.W() / (double) svg.W();
+  double yScale = (double) dest.H() / (double) svg.H();
   double scale = xScale < yScale ? xScale : yScale;
 
   cairo_scale(mContext, scale, scale);
